+++ applications/fishing-map/src/features/editor/editor.utils.ts
@@ -1,0 +1,8 @@
+import { Dataview } from '@globalfishingwatch/api-types/dist'
+
+export const getDataviewInstanceFromDataview = (dataview: Dataview) => {
+  return {
+    id: `${dataview.name}-${Date.now()}`,
+    dataviewId: dataview.id,
+  }
+}