+++ applications/vessel-history/src/features/map/map.slice.ts
@@ -1,6 +1,7 @@
 import { createSlice, PayloadAction, createSelector } from '@reduxjs/toolkit'
 import { Generators } from '@globalfishingwatch/layer-composer'
 import GFWAPI from '@globalfishingwatch/api-client'
+import { ApiEvent } from '@globalfishingwatch/api-types'
 import {
   AnyGeneratorConfig,
   BackgroundGeneratorConfig,
@@ -8,9 +9,12 @@ import {
 } from '@globalfishingwatch/layer-composer/dist/generators/types'
 import { RootState } from 'store'
 import { selectUrlMapZoomQuery, getDateRange } from 'routes/routes.selectors'
+import { Range } from 'types'
 
 export interface MapState {
   generatorsConfig: AnyGeneratorConfig[]
+  highlightedTime?: Range
+  highlightedEvent?: ApiEvent
 }
 
 const initialState: MapState = {
@@ -44,16 +48,28 @@ export const mapSlice = createSlice({
         Object.assign(state.generatorsConfig[index], config)
       }
     },
+    setHighlightedTime: (state, action: PayloadAction<Range>) => {
+      state.highlightedTime = action.payload
+    },
+    setHighlightedEvent: (state, action: PayloadAction<ApiEvent | undefined>) => {
+      state.highlightedEvent = action.payload
+    },
+    disableHighlightedTime: (state) => {
+      state.highlightedTime = undefined
+    },
   },
 })
 
 // createSlice generates for us actions that we can use to modify the store by calling dispatch, or using the useDispatch hook
-export const { updateGenerator } = mapSlice.actions
+export const { updateGenerator, setHighlightedTime, setHighlightedEvent, disableHighlightedTime } =
+  mapSlice.actions
 
 // This is a simple selector that just picks a portion of the stor for consumption by either a component,
 // or a more complex memoized selector. Memoized selectors and/or that need to access several slices,
 // should go into a distiinct [feature].selectors.ts file (use createSelector from RTK)
 export const selectGeneratorsConfig = (state: RootState) => state.map.generatorsConfig
+export const selectHighlightedTime = (state: RootState) => state.map.highlightedTime
+export const selectHighlightedEvent = (state: RootState) => state.map.highlightedEvent
 
 export const selectGlobalGeneratorsConfig = createSelector(
   [selectUrlMapZoomQuery, getDateRange],