+++ packages/layer-composer/src/generators/types.ts
@@ -247,10 +247,17 @@ export interface VesselEventsGeneratorConfig extends GeneratorConfig {
   type: Type.VesselEvents
   data: RawEvent[]
   color?: string
+  event?: {
+    activeIconsSize?: number
+    activeStrokeColor?: string
+    iconsSize?: number
+    strokeColor?: string
+  }
   track?: TrackGeneratorConfigData
   showIcons?: boolean
   showAuthorizationStatus?: boolean
   currentEventId?: string
+  pointsToSegmentsSwitchLevel?: number
 }
 
 /**
@@ -271,9 +278,10 @@ export interface HeatmapGeneratorConfig extends GeneratorConfig {
   // end: string
   // zoom: number
   maxZoom?: number
-  breaks?: number
-  steps?: number[]
+  numBreaks?: number
+  breaks?: number[]
   tilesUrl: string
+  static?: boolean
   statsUrl?: string
   scalePowExponent?: number
   datasets: string[]